requirements.txt
flask==2.3.3
gunicorn==21.2.0
line-bot-sdk==3.4.0
